+++ wm/sampledata/images.py
@@ -1,16 +1,16 @@
 # -*- coding: utf-8 -*-
 from functools import wraps
-from wm.sampledata import logger
-from PIL import Image
 from io import BytesIO
-
+from PIL import Image
+from wm.sampledata import logger
 
 import os
 import random
-import time
-import urllib
 import requests
-
+import time
+import urllib.error
+import urllib.parse
+import urllib.request
 
 
 # see
@@ -35,20 +35,20 @@ def _retry(ExceptionToCheck, tries=4, delay=3, backoff=2, logger=None):
     :param logger: logger to use. If None, print
     :type logger: logging.Logger instance
     """
-    def deco_retry(f):
 
+    def deco_retry(f):
         @wraps(f)
         def f_retry(*args, **kwargs):
             mtries, mdelay = tries, delay
             while mtries > 1:
                 try:
                     return f(*args, **kwargs)
-                except ExceptionToCheck, e:
+                except ExceptionToCheck as e:
                     msg = "%s, Retrying in %d seconds..." % (str(e), mdelay)
                     if logger:
                         logger.warning(msg)
                     else:
-                        print msg
+                        print(msg)
                     time.sleep(mdelay)
                     mtries -= 1
                     mdelay *= backoff
@@ -64,39 +64,50 @@ def _download(url):
     """try to download the url 5 times, with exponential delay.
     first delay 2 seconds, second 4, third 8, etc
     """
-    logger.info('Downloading %s' % url)
-    headers = {'User-Agent': 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:46.0) Gecko/20100101 Firefox/46.0'}  # noqa
+    logger.info("Downloading %s" % url)
+    headers = {
+        "User-Agent": "Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:46.0) Gecko/20100101 Firefox/46.0"
+    }  # noqa
     return requests.get(url, headers=headers).content
 
 
 # Lorempixel.com categories.
 CATEGORIES = [
-    'abstract', 'animals', 'business',
-    'cats', 'city', 'food', 'nightlife',
-    'fashion', 'people', 'nature', 'sports',
-    'technics', 'transport',
+    "abstract",
+    "animals",
+    "business",
+    "cats",
+    "city",
+    "food",
+    "nightlife",
+    "fashion",
+    "people",
+    "nature",
+    "sports",
+    "technics",
+    "transport",
 ]
 
 
-def get_placeholder_image(width=1024, height=768,
-                          text='',  # default to dimensions of image
-                          bkg_color='CCCCCC',
-                          text_color='969696'):
-    """see https://placeholder.com/ for explanations
-    """
+def get_placeholder_image(
+    width=1024,
+    height=768,
+    text="",  # default to dimensions of image
+    bkg_color="CCCCCC",
+    text_color="969696",
+):
+    """see https://placeholder.com/ for explanations"""
 
-    url = u'http://via.placeholder.com/{w}x{h}/{bkg}/{textcolor}'.format(
-        w=width,
-        h=height,
-        bkg=bkg_color.strip('#'),
-        textcolor=text_color.strip('#'))
+    url = "http://via.placeholder.com/{w}x{h}/{bkg}/{textcolor}".format(
+        w=width, h=height, bkg=bkg_color.strip("#"), textcolor=text_color.strip("#")
+    )
 
     params = {}
     if text:
-        params['text'] = text.encode('utf-8')
+        params["text"] = text.encode("utf-8")
 
     if params:
-        url += '?' + urllib.urlencode(params)
+        url += "?" + urllib.parse.urlencode(params)
 
     return _download(url)
 
@@ -112,12 +123,12 @@ def test_stub(f):
         height = 768
         if len(args) >= 2:
             width, height = args[:2]
-        ci = os.environ.get('CI', '0')
-        if ci.lower() in ['1', 'true', 'yes', True]:
+        ci = os.environ.get("CI", "0")
+        if ci.lower() in ["1", "true", "yes", True]:
             logger.info("CI=true set, do not download image")
-            img = Image.new("RGB", (width, height), '#bababa')
+            img = Image.new("RGB", (width, height), "#bababa")
             with BytesIO() as buffer:
-                img.save(buffer, 'JPEG')
+                img.save(buffer, "JPEG")
                 buffer.seek(0)
                 return buffer.read()
         return f(*args, **kwargs)
@@ -126,55 +137,54 @@ def test_stub(f):
 
 
 @test_stub
-def getFlickrImage(width=1024, height=768,
-                   keywords=[], match_all_keywords=False,
-                   gray=False):
+def getFlickrImage(
+    width=1024, height=768, keywords=[], match_all_keywords=True, gray=False
+):
     """obtains an image from loremflickr.com.
 
-    If you set match_all_keywords to true it will
-    search for all keywords (AND).
+    If you set match_all_keywords to false it will
+    search for any of the keywords (OR).
 
     Set gray to True to get a grayscale image.
     """
 
-    url = 'http://loremflickr.com/'
+    url = "http://loremflickr.com/"
     params = []
 
     if gray:
-        params.append('g')
+        params.append("g")
 
-    params.append('%d/%d' % (width, height))
-    params.append(','.join(keywords))
+    params.append("%d/%d" % (width, height))
+    params.append(",".join(keywords or []))
     if match_all_keywords:
-        params.append('all')
+        params.append("all")
 
-    url = url + '/'.join(params)
+    url = url + "/".join(params)
 
     return _download(url)
 
 
 @test_stub
-def getImage(width=1024, height=768, category=None,
-             gray=False, index=None, text=None):
+def getImage(width=1024, height=768, category=None, gray=False, index=None, text=None):
     """obtains an image from lorempixel.com
 
     for possible categories see ``CATEGORIES``.
     """
 
-    url = 'http://lorempixel.com/'
+    url = "http://lorempixel.com/"
     params = []
 
     if gray:
-        params.append('g/')
-    params.append('%d/%d/' % (width, height))
+        params.append("g/")
+    params.append("%d/%d/" % (width, height))
     if category:
-        params.append(category + '/')
+        params.append(category + "/")
         if index:
-            params.append('%d/' % index)
+            params.append("%d/" % index)
     if text:
-        params.append('%s/' % text)
+        params.append("%s/" % text)
 
-    url = url + ''.join(params)
+    url = url + "".join(params)
 
     return _download(url)
 
@@ -182,9 +192,14 @@ def getImage(width=1024, height=768, category=None,
 RATIOS = [(16, 9), (4, 3)]
 
 
-def getRandomFlickrImage(long_edge=1024,
-                         keywords=[], match_all_keywords=False,
-                         gray=None, landscape=None, ratios=RATIOS):
+def getRandomFlickrImage(
+    long_edge=1024,
+    keywords=[],
+    match_all_keywords=False,
+    gray=None,
+    landscape=None,
+    ratios=RATIOS,
+):
     """Returns a random image from loremflickr.com (portrait or landscape)
     in one of the available RATIOS.
 
@@ -208,35 +223,39 @@ def getRandomFlickrImage(long_edge=1024,
         width = int(long_edge / ratio)
 
     return getFlickrImage(
-        width, height,
+        width,
+        height,
         keywords=keywords,
         match_all_keywords=match_all_keywords,
-        gray=gray
+        gray=gray,
     )
 
-
-def getRandomImage(long_edge=1024, category=None, gray=None,
-                   landscape=None, ratios=RATIOS):
-    """returns a random image from lorempixel.com (portrait or landscape)
-    in one of the available RATIOS
-
-    set gray to True to force grayscale pictures and to False to force color
-    pictures
-    """
+def random_dimensions(long_edge=1024, landscape=None, ratios=RATIOS):
+    """returns (width, height) for random ratio and random orientation"""
     ratio = random.choice(ratios)
     ratio = float(ratio[0]) / ratio[1]
-
     if landscape is None:
         landscape = random.choice([True, False])
-
-    if gray is None:
-        gray = random.choice([True, False])
-
     if landscape:
         width = long_edge
         height = int(long_edge / ratio)
     else:
         height = long_edge
         width = int(long_edge / ratio)
+    return width, height
+
+def getRandomImage(
+    long_edge=1024, category=None, gray=None, landscape=None, ratios=RATIOS
+):
+    """returns a random image from lorempixel.com (portrait or landscape)
+    in one of the available RATIOS
+
+    set gray to True to force grayscale pictures and to False to force color
+    pictures
+    """
+    width, height = random_dimensions(long_edge, landscape, ratios)
+
+    if gray is None:
+        gray = random.choice([True, False])
 
     return getImage(width, height, category=category, gray=gray)